+++ components/JobCard.tsx
@@ -1,7 +1,7 @@
 // src/components/JobCard.tsx
 import React from 'react';
 import { View, Text, StyleSheet, TouchableOpacity } from 'react-native';
-import { Job } from '../api/jobApi';
+import { Job } from '../api/jobApi(HARDCODEDTESTDATA)';
 
 type JobCardProps = {
   job: Job;
